+++ CHANGELOG.md
@@ -20,6 +20,7 @@ _This project follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) a
 - Fixed potential error while aggregating data from different brokers
   - Fixed error aggregating numeric types
   - Fixed error sorting dates
+- DEGIRO: Prevent authenticated users from being redirected to 2FA login screen
 
 ### Security